+++ test/str.test.js
@@ -164,3 +164,18 @@ describe('slice2()', function () {
   });
 
 });
+
+
+describe('repeat()', function () {
+
+  it('should return a new string, containing specified number of copies of initial string', function () {
+
+    const repeatFoo = _.repeat('foo');;
+    const repeatSpace = _.repeat(' ');
+
+    expect(repeatFoo(3)).to.equal('foofoofoo');
+    expect(repeatSpace(5)).to.equal('     ');
+
+  });
+
+});